requirements.txt
aiohttp==2.3.7
astroid==1.6.0
async-timeout==2.0.0
attrs==17.4.0
autopep8==1.3.3
chardet==3.0.4
idna==2.6
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
multidict==3.3.2
pkg-resources==0.0.0
pycodestyle==2.3.1
pylint==1.8.1
PyYAML==3.12
six==1.11.0
voluptuous==0.10.5
wrapt==1.10.11
yarl==0.18.0
